[#28604703] Upgrade to FactoryGirl
Upgraded FactoryGirl and changed specs to use the new syntax returning
to green.

Changed the style of the TwitterAccount spec to use subject() as an
example of how a spec could be written without instance variables. Only
changed this one small spec to not get ahead of myself, as the others
work just fine.

Using Randome.rand() in the Customer FactoryGirl definition for 1.9.x

Removed references to Machinist and it's blueprint.rb

Added the formtastic form.action wrapping the submit buttons.

Added guard and a Guardfile if one is inclined to use it.